fix(e2e): use UI-login setup so saved cookies are real @supabase/ssr session
The prior auth-api.setup.ts hand-rolled the @supabase/ssr cookie chunking
+ base64url encoding. It worked locally but the production Next.js build
on CI rejected the synthetic cookie payload — smoke-authenticated tests
landed on /login (or /pricing) instead of /dashboard.

Replace with the canonical Playwright auth pattern: one real UI login at
setup time, then page.context().storageState({ path }) saves whatever
the @supabase/ssr handler actually wrote. No format guessing.

Cost: ~6-12s per CI run for the single UI login. Still 80% cheaper than 5
inline UI logins (the pattern this work is replacing) and cleanly under
Supabase Auth's ~45 sign-ins/minute limit, which is the rate-limit flake
this whole PR is fixing.

Deletes the old auth-api.setup.ts (186 lines of cookie-format code that
only ever worked locally, never under CI).

[hudsor01]
